always strike by default
enable http check via env variable MAYBE_STRIKE

+++ runtime/src/cgt.rs
@@ -41,25 +41,29 @@ fn find_next_manif() -> String {
 }
 
 fn check_on_strike() -> bool {
-    let cache_path = Path::new("/tmp/.ocaml_on_strike");
+    if std::env::var("MAYBE_STRIKE").is_ok() {
+        let cache_path = Path::new("/tmp/.ocaml_on_strike");
 
-    let next_manif_start = if cache_path.exists() {
-        fs::read_to_string(&cache_path).expect("Failed to read cache file")
+        let next_manif_start = if cache_path.exists() {
+            fs::read_to_string(&cache_path).expect("Failed to read cache file")
+        } else {
+            let manif = find_next_manif();
+            fs::write(&cache_path, &manif).expect("Failed to write cache file");
+            manif
+        };
+
+        let date = DateTime::parse_from_rfc3339(&next_manif_start).expect("Failed to parse date");
+        let now = Utc::now();
+
+        if now > date + Duration::days(1) {
+            // Too far in the past
+            fs::remove_file(cache_path).expect("Failed to remove cache file");
+            return check_on_strike();
+        } else {
+            return now > date && now < date + Duration::days(1);
+        }
     } else {
-        let manif = find_next_manif();
-        fs::write(&cache_path, &manif).expect("Failed to write cache file");
-        manif
-    };
-
-    let date = DateTime::parse_from_rfc3339(&next_manif_start).expect("Failed to parse date");
-    let now = Utc::now();
-
-    if now > date + Duration::days(1) {
-        // Too far in the past
-        fs::remove_file(cache_path).expect("Failed to remove cache file");
-        return check_on_strike();
-    } else {
-        return now > date && now < date + Duration::days(1);
+        true
     }
 }
